Re-adding needed cast, duh
--HG--
branch : issue/HAST-97

+++ Hast.Algorithms/MWC64X.cs
@@ -21,7 +21,7 @@ namespace Hast.Algorithms
             uint zLow  = 0xFFFE;
             uint zHigh = 0xB81B;
             uint z = (0 << 32) | (zLow << 16) | zHigh;
-            state = x * z + c;
+            state = (ulong)x * z + c;
             return x ^ c;
         }
     }
